\name{MuPapers-package}
\alias{MuPapers-package}
\alias{MuPapers}
\docType{package}
\title{
\packageTitle{MuPapers}
}
\description{
\packageDescription{MuPapers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MuPapers}
\packageIndices{MuPapers}
}
\author{
\packageAuthor{MuPapers}

Maintainer: \packageMaintainer{MuPapers}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
