\name{zhang2013operator}
\alias{zhang2013operator}
\docType{data}
\title{
zhang2013operator
}
\description{
zhang2013operator
}
\usage{data("zhang2013operator")}
\format{
  A data frame with 11 observations on the following 7 variables.
  \describe{
    \item{\code{program}}{a factor with levels \code{banker} \code{barbecue} \code{barbecue1.5} \code{bub} \code{cal} \code{commonslang} \code{euclid} \code{find} \code{insert} \code{jaxen} \code{jdepend} \code{jgrapht0.5.0} \code{jgrapht0.5.1} \code{jgrapht0.5.2} \code{jgrapht0.5.3} \code{jgrapht0.6.0} \code{jgrapht0.7.1} \code{jgrapht0.7.3} \code{jmeter} \code{jodatime} \code{jtopas} \code{mid} \code{mime4J} \code{printtokens} \code{printtokens2} \code{quad} \code{replace} \code{schedule} \code{schedule2} \code{tcas} \code{timeandmoney2} \code{timeandmoney3} \code{timemoney} \code{totinfo} \code{trityp} \code{warshal} \code{xmlsecurity} \code{xmlsecurity1.2.1} \code{xmlsecurity1.3.0} \code{xstream}}
    \item{\code{nloc}}{a numeric vector}
    \item{\code{tests}}{a numeric vector}
    \item{\code{nm}}{a numeric vector}
    \item{\code{neq}}{a numeric vector}
    \item{\code{paper}}{a factor with levels \code{ammann2014establishing} \code{offutt1996an} \code{siami2008sufficient} \code{sridharan2010prioritizing} \code{zhang2010isoperator} \code{zhang2013operator} \code{zhang2014empirical}}
    \item{\code{muscore}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
data(zhang2013operator)
## maybe str(zhang2013operator) ; plot(zhang2013operator) ...
}
\keyword{datasets}
